Add `RenderHooks` field to `Config`
This struct will remember which markdown constructs the user wants to
have their rendering overriden.

NOTE: It's being implemented as a struct where each field corresponds to
a construct, using `Option::None` to mark those that won't be overriden,
but we can also use a `HashMap`. It should be best to use a struct because
we don't have very many different constructs to allow overrides.